fix: mapping error

+++ DA_Models/CharacterModels/CharacterDTO.cs
@@ -51,7 +51,7 @@ namespace DA_Models.CharacterModels
             }
             get =>  new DateModel(DateNumber);
         }
-        public int Gold { get; set; } = 0;
+        public decimal Gold { get; set; } = 0.0m;
         public override string ToString() => NPCName;
 
     }
